graph, graphql, store: Fix ParentLink entries
We used to omit entries for parents that have no children from the
`ParentLink` which would violate the constraint that the child/children for
the ith parent is in the ith place in the list of child ids.

That is not a problem for the SQL queries we generate, but can be an issue
for other uses of the data structure.